+++ anyblok_wms_base/bloks/wms_core/__init__.py
@@ -17,7 +17,7 @@ class WmsCore(Blok):
 
     @classmethod
     def import_declaration_module(cls):
-        from . import ns  # noqa
+        from . import wms  # noqa
         from . import location  # noqa
         from . import operation  # noqa
         from . import goods  # noqa