+++ anyblok_wms_base/bloks/wms_core/location.py
@@ -6,13 +6,15 @@
 # This Source Code Form is subject to the terms of the Mozilla Public License,
 # v. 2.0. If a copy of the MPL was not distributed with this file,You can
 # obtain one at http://mozilla.org/MPL/2.0/.
-from sqlalchemy import or_
+from sqlalchemy import orm
+from sqlalchemy import func
+from sqlalchemy import literal
+
 from anyblok import Declarations
 from anyblok.column import String
+from anyblok.column import Text
 from anyblok.column import Integer
 from anyblok.relationship import Many2One
-
-from anyblok_wms_base.constants import DATE_TIME_INFINITY
 
 register = Declarations.register
 Model = Declarations.Model
@@ -22,6 +24,20 @@ Model = Declarations.Model
 class Location:
     """A stock location.
 
+    Locations are perhaps ill-named (see :ref:`improvement_location_name`), but
+    ultimately they represent the idea of "where" the Goods are. There's no
+    reason a priori to consider that Locations themselves are not moving.
+
+    Locations form a tree-like structure (a forest), which is encoded through
+    the :attr:`parent` field, and have :attr:`tags <tag>` which can be used to
+    express functional meaning, especially for
+    :meth:`quantity computations
+    <anyblok_wms_base.bloks.wms_core.wms.Wms.quantity>`.
+
+    Downstream libraries and applications which don't want to use this hierarchy
+    and the defaulting of tags that comes along can do so by overriding
+    :meth:`flatten_hierarchy_with_tags` and :meth:`resolve_tag`
+
     TODO add location types to encode behavioral properties (internal, EDI,
     stuff like size ?)
     """
@@ -30,6 +46,27 @@ class Location:
     label = String(label="Label")
     parent = Many2One(label="Parent location",
                       model='Model.Wms.Location')
+    tag = Text()
+    """Tag for Quantity grouping.
+
+    This field is a kind of tag that can be used to filter in quantity
+    queries. It allows for location-based assessment of stock levels by
+    recursing in the hierarchy while still allowing exceptions: to discard or
+    include sublocations.
+
+    For instance, one may represent a big warehouse as having several rooms
+    (R1, R2)
+    each one with an examination area (R1/QA, R2/QA), which can be further
+    subdivided.
+
+    Goods stored in the workshop are not to be sold, except maybe those that
+    are in the waiting area before been put back in stock (R1/QA/Good etc.).
+
+    It's then useful to tag Rooms as 'sellable', but in them, QA locations as
+    'qa', and finally the good waiting areas as 'sellable' again.
+
+    See in unit tests for a demonstration of that.
+    """
 
     def __str__(self):
         return ("(id={self.id}, code={self.code!r}, "
@@ -38,73 +75,91 @@ class Location:
     def __repr__(self):
         return "Wms.Location" + str(self)
 
-    def quantity(self, goods_type, additional_states=None, at_datetime=None):
-        """Return the full quantity in location for the given type.
+    def resolve_tag(self):
+        """Return self.tag, or by default ancestor's."""
+        # TODO make a recursive query for this also
+        if self.tag is not None:
+            return self.tag
+        if self.parent is None:
+            return None
+        return self.parent.resolve_tag()
 
-        :param additional_states:
-            Optionally, states of the Goods Avatar to take into account
-            in addition to the ``present`` state.
+    def quantity(self, goods_type, recursive=True, **kwargs):
+        """Goods quantity computation within the current location.
 
-            Hence, for ``additional_states=['past']``, we have the
-            Goods Avatars that were already there and still are,
-            as well as those that aren't there any more,
-            and similarly for 'future'.
-        :param at_datetime:
-            take only into account Goods Avatar whose date-time range
-            contains the specified value.
+        This is kept for backwards compatibility and simply calls
+        :meth:`Wms.quantity()
+        <anyblok_wms_base.bloks.wms_core.wms.Wms.quantity>`
 
-            ``anyblok_wms_base.constants.DATE_TIME_INFINITY``
-            can in particular be used to consider only those
-            Avatars whose ``dt_until`` is ``None``.
-
-            This parameter is mandatory if ``additional_states`` is specified.
-
-        TODO: make recursive (not fully decided about the forest structure
-        of locations)
-
-        TODO: provide filtering according to Goods properties (should become
-        special PostgreSQL JSON clauses)
-
-        TODO PERF: for timestamp ranges, use GiST indexes and the @> operator.
-        See the comprehensive answer to `that question
-        <https://dba.stackexchange.com/questions/39589>`_ for an entry point.
-        Let's get a DB with serious volume and datetimes first.
+        (see source code for details on arguments passing)
         """
-        Goods = self.registry.Wms.Goods
-        Avatar = Goods.Avatar
-        query, use_count = self.base_quantity_query()
-        query = query.filter(Goods.type == goods_type, Avatar.location == self)
-
-        if additional_states is None:
-            query = query.filter(Avatar.state == 'present')
-        else:
-            states = ('present',) + tuple(additional_states)
-            query = query.filter(Avatar.state.in_(states))
-            if at_datetime is None:
-                raise ValueError(
-                    "Querying quantities with additional states {!r} requires "
-                    "to specify the 'at_datetime' kwarg".format(
-                        additional_states))
-
-        if at_datetime is DATE_TIME_INFINITY:
-            query = query.filter(Avatar.dt_until.is_(None))
-        elif at_datetime is not None:
-            query = query.filter(Avatar.dt_from <= at_datetime,
-                                 or_(Avatar.dt_until.is_(None),
-                                     Avatar.dt_until > at_datetime))
-        if use_count:
-            return query.count()
-
-        res = query.one()[0]
-        return 0 if res is None else res
+        return self.registry.Wms.quantity(location=self,
+                                          location_recurse=recursive,
+                                          goods_type=goods_type,
+                                          **kwargs)
 
     @classmethod
-    def base_quantity_query(cls):
-        """Return base join query, without any filtering, and eval indication.
+    def flatten_subquery_with_tags(cls, top=None, resolve_top_tag=True):
+        """Return an SQL subquery flattening the hierarchy, resolving tags.
 
-        :return: query, ``True`` if ``count()`` is to be used. Otherwise,
-                 the query is assumed to produce exactly one row, with the
-                 wished quantity result (possibly ``None`` for 0)
+        The resolving tag policy is that a Location whose tag is ``None``
+        inherits its parent's.
+
+        This subquery cannot be used directly: it is meant to be used as part
+        of a wider query; see unit tests (``test_location``) for nice examples
+        with or without joins. It has two columns: ``id`` and ``tag``.
+
+        :param top:
+           if specified, the query starts at this Location (inclusive)
+
+        This default implementation issues a recursive CTE (``WITH RECURSIVE``)
+        that climbs down along the :attr:`parent` field.
+
+        For some applications with a large and complicated Location hierarchy,
+        joining on this CTE can become a performance problem. Quoting
+        `PostgreSQL documentation on CTEs
+        <https://www.postgresql.org/docs/10/static/queries-with.html>`_:
+
+          However, the other side of this coin is that the optimizer is less
+          able to push restrictions from the parent query down into a WITH
+          query than an ordinary subquery.
+          The WITH query will generally be evaluated as written,
+          without suppression of rows that the parent query might
+          discard afterwards.
+
+        If that becomes a problem, it is still possible to override the
+        present method: any subquery whose results have the same columbs
+        can be used by callers instead of the recursive CTE.
+
+        Examples:
+
+        1. one might design a flat Location hierarchy using prefixing on
+           :attr:`code` to express inclusion instead of the provided
+           :attr:`parent`. See :meth:`anyblok_wms_base.bloks.wms_core.tests
+           .test_location.test_override_tag_recursion` for a proof of concept
+           of this.
+        2. one might make a materialized view out of the present recursive CTE,
+           refreshing as soon as needed.
         """
-        Avatar = cls.registry.Wms.Goods.Avatar
-        return Avatar.query().join(Avatar.goods), True
+        query = cls.registry.session.query
+
+        if top is not None and top.tag is None:
+            init_tag = top.resolve_tag()
+            cte = cls.query(cls.id, literal(init_tag).label('tag'))
+        else:
+            cte = cls.query(cls.id, cls.tag)
+        if top is None:
+            cte = cte.filter(cls.parent == top)  # doesn't work with is_()
+        else:
+            # starting with top, not its children so that the children
+            # can inherit tag from top (done in the recursive part of the
+            # subquery)
+            cte = cte.filter(cls.id == top.id)
+        cte = cte.cte(name="location_tag", recursive=True)
+        parent = orm.aliased(cte, name='parent')
+        child = orm.aliased(cls, name='child')
+        cte = cte.union_all(
+            query(child.id,
+                  func.coalesce(child.tag, parent.c.tag).label('tag')).filter(
+                      child.parent_id == parent.c.id))
+        return cte